// src/shared/api/models/UserRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserRead = {
    id: string;
    created: string;
    updated: string;
    auth_id: string;
    email: string;
    username: string;
    picture: string;
};